Pin SciPy version to latest; regenerate `scipy.special` signatures
Closes https://github.com/numba/numba-scipy/issues/35.

At this early point in the project I don't think it's worth trying to
support multiple versions of SciPy simultaneously; this could easily
overwhelm limited developer time better spent adding new
features. Instead this pins SciPy at the latest version (1.4.1), the
idea being that `numba-scipy` and `scipy` releases are done in
lockstep.

It's easy to relax this in the future if more resources are available.