fix the label for uploading tags
This fixes the upload of tags to `numba/label/main` channel on
anaconda.org. Previously, tags would be uploaded to:
`numba/label/numba/label/dev` which would show up in the webinterface as
`numba/label/dev` and not just `dev`. But, we also want tags to arive on
`numba/label/main` so we use `main` as argument for the final upload
command, which will be along the lines of:

```
anaconda upload -u numba -l main ...
```